Introduce modular reactive values/functions concept
Because we decided to split the server and UI definitions into
separate "modules", we also needed to create a way for different
modules to exchange their reactive values and functions.

In this regard, each module is going to have two server-like
functions, react_module-name_f and react_module-name_v, which will
define and return reactive functions and objects, respectively. Each
such function is server-like, i.e. it will take input, output and
session arguments. The two remaining arguments are rv and rf, for the
input reactive values and functions.

* R/shiny-ui-config.R(react_conf_v): New function.
  Add conf-related reactive values.

  (react_vonf_f): New function. Add conf-related reactive functions.

  (server_conf): Update observers.

* R/shiny-ui-top.R(server): Adapt to the new concept.