server_conf: Add new observers
In the current spirit of naming the observers in the comments, we are
listing the changes under their names inside the contex (angle)
brackets.

Note that some of the observers which were listed as "updated" may, or
may have not been created right now. It was too much to track.

* R/shiny-ui-config.R(server_conf) <update-files-on-restore,update-files-on-mzmlb>:
  New observers. They finally complete the data files modification,
  addition mechanism.

(server_conf) <build-config>: Updated.

(server_conf) <build-compounds>: Updated.