shiny-ui-config: Implement reactive state build-up
From the beginning, the intention was to follow the simple signal
passing paradigm, where a single state object is being passed from
function to function. With a GUI this is complicated by continuously
changing state and in a reactive paradigm of Shiny, additional
complication arises from inability to distinguish between different
inputs for the same destination observer, or reactive function.

But, now that the problem has been solved by adding new reactive state
variables for the project path and the data file control, it was
possible to emulate the streamlined signal processing flow by adding a
few new reactive functions.

Here they are.

* R/shiny-ui-config.R(m_conf,m_input,m): New reactive functions. Each
  builds its part of the m object. Each calls the preceeding one.