api, mix: Make configuration static
Prior to this commit, cofiguration consisted of static values and
dynamic variables calculated based on the static values. Changing a
static value meant that those variables had to be recalculated. In a
reactive environment where things change a lot, there was danger that
some parts do not get calculated, or that recalculation leads to
overwriting of other variables. Now, all the dynamic config entries,
such as paths within the project tree are delegated to "accessor"
methods that evaluate their value at the right time.

* R/mix(get_fn_comp,get_fn_extr,get_fn_summ): New functions to
evaluate paths to the comprehensive table, extraction file and the
summary file.

* R/api.R: Corresponding changes in here.

* NAMESPACE: Some updates.